Enabled source map generation when compiling the TypeScript files.
The source map improves the quality of in-browser script debugging.
This required a movement of all the TypeScript source files to the
static files directory so they can be copied to the output.
Also modified the measurement filtering so that the ToTime parameter
now has priority over the TimePeriod parameter when both parameters
are provided simultaneously.