Added the latest data page to the monitor.
On this page, the latest logged data entry is displayed. Now it is used
as a main page instead of the last day data chart page.
Added a warning indicating that the latest data have been logged more
that N minutes ago to indicate possible data logging failures. The time
period can be customized through configuration.
Also added a signed in user name to the navigation menu.