Added the BmeReaderProvider measurement provider.
This provider interacts with the BMEReader adapter device that
represents a USB to I2C bus hardware bridge between the computer and
the BME280 sensor connected to it. The device is registered in system
as a virtual serial port, and it can be accessed by its serial port
name that must be provided in the configuration file.